% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonApplyTermWeightFunctions}
\alias{LSAMonApplyTermWeightFunctions}
\title{Apply term-weight functions.}
\usage{
LSAMonApplyTermWeightFunctions(lsaObj, globalWeightFunction = "Entropy",
  localWeightFunction = "None", normalizerFunction = "Cosine")
}
\arguments{
\item{lsaObj}{A LSAMon object.}

\item{globalWeightFunction}{Global weight function.}

\item{localWeightFunction}{Local weight function.}

\item{normalizerFunction}{Normalizer function.}
}
\value{
A LSAMon object.
}
\description{
Re-weight the entries of the document-term matrix.
}
\details{
This function simply calls
\code{\link{SparseMatrixRecommender::SMRApplyTermWeightFunctions}} for the
document-term matrix \code{lsaObj$DocumentTermMatrix}.
The obtained matrix is assigned to \code{lsaObj$WeightedDocumentTermMatrix}.
}
