% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonBasisVectorInterpretation}
\alias{LSAMonBasisVectorInterpretation}
\title{Basis vectors interpretation.}
\usage{
LSAMonBasisVectorInterpretation(lsaObj, vectorIndices = NULL, n = 12,
  orderBySignificanceQ = TRUE)
}
\arguments{
\item{lsaObj}{A LSAMon object.}

\item{n}{Number of (top) coordinates per basis vector.}

\item{orderBySignificanceQ}{Should the basis vectors be ordered by their significance?}

\item{basisVectorIndexes}{Basis vectors to be interpretted.
If NULL all indexes are taken.}
}
\value{
A LSAMon object.
}
\description{
Interpret specified basis vectors.
}
