% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonEchoDocumentTermMatrixStatistics}
\alias{LSAMonEchoDocumentTermMatrixStatistics}
\title{Echo document-term matrix statistics.}
\usage{
LSAMonEchoDocumentTermMatrixStatistics(lsaObj, logBase = NULL, echoQ = T)
}
\arguments{
\item{lsaObj}{A LSAMon object.}

\item{logBase}{Logarithmic base for the document-term matrix row and column sums.}

\item{echoQ}{Should the result be echoed or not?}
}
\value{
A LSAMon object.
}
\description{
Echoes a table and histograms of document-term matrix statistics.
}
