% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonEchoFunctionValue}
\alias{LSAMonEchoFunctionValue}
\title{Function application to monad's value.}
\usage{
LSAMonEchoFunctionValue(lsaObj, f)
}
\arguments{
\item{lsaObj}{An LSAMon object.}

\item{f}{A function to be applied to \code{lsaObj$Value}.}
}
\value{
A LSAMon object.
}
\description{
Apply a function to the "Value" element/member of the monad object.
}
\details{
Prints \code{f(lsaObj$Value)}.
}
