% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonEchoValue}
\alias{LSAMonEchoValue}
\title{Print monad's value.}
\usage{
LSAMonEchoValue(lsaObj)
}
\arguments{
\item{lsaObj}{An LSAMon object.}
}
\value{
A LSAMon object.
}
\description{
Prints the"Value" element/member of the monad object.
}
\details{
Prints \code{flsaObj$Value}.
}
