% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonExportToDirectory}
\alias{LSAMonExportToDirectory}
\title{Export to a directory}
\usage{
LSAMonExportToDirectory(
  lsaObj,
  directoryPath,
  dataNameInfix = "",
  csvTripletsQ = FALSE,
  digits = NULL
)
}
\arguments{
\item{lsaObj}{An LSAMon object}

\item{directoryPath}{A path to a directory.}

\item{dataNameInfix}{A string infix designating the exported LSA object.}

\item{csvTripletsQ}{Should the LSA matrix be also exported to a CSV file.}

\item{digits}{An integer for the number of digits to round to.
See \code{\link{round}}.
If NULL no rounding is done.
Only applies to the file with the name pattern "SMR-M01-from-*.csv".}
}
\value{
A LSAMon object
}
\description{
Exports the LSAMon to specified directory.
}
\details{

}
