% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\docType{data}
\name{LSAMonFailureSymbol}
\alias{LSAMonFailureSymbol}
\title{Failure symbol for LSAMon.}
\format{An object of class \code{logical} of length 1.}
\usage{
LSAMonFailureSymbol
}
\description{
Failure symbol for the monad LSAMon.
}
\keyword{datasets}
