% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonFindMostImportantTexts}
\alias{LSAMonFindMostImportantTexts}
\title{Fins most important texts.}
\usage{
LSAMonFindMostImportantTexts(lsaObj, nTop = 5)
}
\arguments{
\item{lsaObj}{A LSAMon object}

\item{nTop}{The number of top most important texts.}
}
\value{
A LSAMon object
}
\description{
Finds the most important texts in the monad.
}
\details{
The result is a data frame with columns \code{c("Score", "ID")};
it is assigned to \code{lsaObj$Value}.
}
