% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonGetStatisticalThesaurus}
\alias{LSAMonGetStatisticalThesaurus}
\title{Get statistical thesaurus table.}
\usage{
LSAMonGetStatisticalThesaurus(
  lsaObj,
  words,
  numberOfNearestNeighbors = 12,
  wideFormQ = FALSE,
  echoQ = FALSE
)
}
\arguments{
\item{lsaObj}{A LSAMon object.}

\item{words}{Words to find thesaurus entries for.}

\item{numberOfNearestNeighbors}{Number of nearest neighbors per specified word.}

\item{wideFormQ}{Should the thesaurus table be in wide form or not?}

\item{echoQ}{Should the thesaurus table be echoed or not?}
}
\value{
A LSAMon object.
}
\description{
Gets a table of statistical thesaurus entries for specified words.
}
