% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonImportFromDirectory}
\alias{LSAMonImportFromDirectory}
\title{Import from a directory}
\usage{
LSAMonImportFromDirectory(
  lsaObj,
  directoryPath,
  dataNamePrefix = "",
  dataNameInfix = "",
  format = NULL
)
}
\arguments{
\item{lsaObj}{An LSAMon object}

\item{directoryPath}{A path to a directory.}

\item{dataNamePrefix}{A string prefix designating the exported LSA object.}

\item{dataNameInfix}{A string infix designating the exported LSA object.}

\item{format}{Export format, one of "CSV", "CSVHarwellBoeing", "feather", or NULL.
If NULL then "CSVHarwellBoeing" is used.}

\item{digits}{An integer for the number of digits to round to.
See \code{\link{round}}.
If NULL no rounding is done.
Only applies to the file with the name pattern "SMR-M01-from-*.csv".}
}
\value{
A LSAMon object
}
\description{
Imports a LSAMon from a specified directory.
}
\details{

}
