% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonMakeDocumentTermMatrix}
\alias{LSAMonMakeDocumentTermMatrix}
\title{Document-term matrix creation.}
\usage{
LSAMonMakeDocumentTermMatrix(lsaObj, splitPattern = "\\\\W",
  stemWordsQ = FALSE, stopWords = NULL)
}
\arguments{
\item{lsaObj}{A LSAMon object.}

\item{splitPattern}{Pattern to use for splitting the documents into words.}

\item{stemWordsQ}{Should the words be stemmed or not?}

\item{stopWords}{A character vector with stop words to be removed.}
}
\value{
A LSAMon object.
}
\description{
Make a document-term matrix for a collection of documents
using specified word splitting and optional word stemming.
}
\details{
The obtained sparse matrix is \code{lsaObj$DocumentTermMatrix}.
The documents are expected to have ID's. If \code{is.null(names(lsaObj$Documents))}
then the ID's are just the indexes of the documents list/vector.
}
