% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonMakeTopicsTable}
\alias{LSAMonMakeTopicsTable}
\title{Make topics table.}
\usage{
LSAMonMakeTopicsTable(
  lsaObj,
  numberOfTerms = 12,
  wideFormQ = FALSE,
  numberOfTableColumns = NA
)
}
\arguments{
\item{lsaObj}{A LSAMon object.}

\item{numberOfTerms}{Number of terms per topic.}

\item{wideFormQ}{Should the topics table be in wide form or not?}

\item{numberOfTableColumns}{Number of table columns.
(Dummy argument at this point.)}
}
\value{
A LSAMon object.
}
\description{
Makes a table of (the calculated) topics.
}
