% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonNormalizeMatrixProduct}
\alias{LSAMonNormalizeMatrixProduct}
\title{Normalized matrix product components.}
\usage{
LSAMonNormalizeMatrixProduct(
  lsaObj,
  normalizeLeftQ = TRUE,
  orderBySignificanceQ = FALSE
)
}
\arguments{
\item{lsaObj}{A LSAMon object.}

\item{normalizeLeftQ}{Should the left factor be normalized?}

\item{orderBySignificanceQ}{Should the basis vectors be ordered by their significance?}
}
\value{
A LSAMon object.
}
\description{
Normalizes the matrix factors and returns a list with them.
}
