% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonRepresentByTopics}
\alias{LSAMonRepresentByTopics}
\title{Represent by topics.}
\usage{
LSAMonRepresentByTopics(lsaObj, query, applyTermWeightFunctionsQ = TRUE)
}
\arguments{
\item{lsaObj}{A LSAMon object.}

\item{query}{A character vector or a sparse matrix to be represented
in the space of monad's document-term matrix.}

\item{applyTermWeightFunctionsQ}{Should the weight term functions
be applied to the result matrix or not?}
}
\value{
A LSAMon object.
}
\description{
Find the topics representation of a matrix or a document.
}
\details{
If the argument \code{query} is a character then this function
applies first the function \code{\link{LSAMonRepresentByTerms}}.
}
