% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonSetAutomaticTopicNames}
\alias{LSAMonSetAutomaticTopicNames}
\title{Set AutomaticTopicNames.}
\usage{
LSAMonSetAutomaticTopicNames(lsaObj, AutomaticTopicNames)
}
\arguments{
\item{lsaObj}{An LSAMon object.}

\item{AutomaticTopicNames}{A list of regression objects.}
}
\value{
An LSAMon object.
}
\description{
Sets AutomaticTopicNames into the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{LSAMonSetDocumentTermMatrix}},
  \code{\link{LSAMonSetDocuments}},
  \code{\link{LSAMonSetH}}, \code{\link{LSAMonSetTerms}},
  \code{\link{LSAMonSetTopicColumnPositions}},
  \code{\link{LSAMonSetValue}},
  \code{\link{LSAMonSetWeightedDocumentTermMatrix}},
  \code{\link{LSAMonSetW}},
  \code{\link{LSAMonTakeAutomaticTopicNames}},
  \code{\link{LSAMonTakeDocumentTermMatrix}},
  \code{\link{LSAMonTakeDocuments}},
  \code{\link{LSAMonTakeH}}, \code{\link{LSAMonTakeTerms}},
  \code{\link{LSAMonTakeTopicColumnPositions}},
  \code{\link{LSAMonTakeValue}},
  \code{\link{LSAMonTakeWeightedDocumentTermMatrix}},
  \code{\link{LSAMonTakeW}}
}
\concept{Set/Take functions}
