% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonSetDocuments}
\alias{LSAMonSetDocuments}
\title{Set Documents.}
\usage{
LSAMonSetDocuments(lsaObj, Documents)
}
\arguments{
\item{lsaObj}{An LSAMon object.}

\item{Documents}{A list of regression objects.}
}
\value{
An LSAMon object.
}
\description{
Sets Documents into the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{LSAMonSetAutomaticTopicNames}},
  \code{\link{LSAMonSetDocumentTermMatrix}},
  \code{\link{LSAMonSetH}}, \code{\link{LSAMonSetTerms}},
  \code{\link{LSAMonSetTopicColumnPositions}},
  \code{\link{LSAMonSetValue}},
  \code{\link{LSAMonSetWeightedDocumentTermMatrix}},
  \code{\link{LSAMonSetW}},
  \code{\link{LSAMonTakeAutomaticTopicNames}},
  \code{\link{LSAMonTakeDocumentTermMatrix}},
  \code{\link{LSAMonTakeDocuments}},
  \code{\link{LSAMonTakeH}}, \code{\link{LSAMonTakeTerms}},
  \code{\link{LSAMonTakeTopicColumnPositions}},
  \code{\link{LSAMonTakeValue}},
  \code{\link{LSAMonTakeWeightedDocumentTermMatrix}},
  \code{\link{LSAMonTakeW}}
}
\concept{Set/Take functions}
