% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonSetTerms}
\alias{LSAMonSetTerms}
\title{Set Terms.}
\usage{
LSAMonSetTerms(lsaObj, terms)
}
\arguments{
\item{lsaObj}{An LSAMon object.}

\item{terms}{A list of terms.}
}
\value{
An LSAMon object.
}
\description{
Sets Terms into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{LSAMonSetDocumentTermMatrix}()},
\code{\link{LSAMonSetDocuments}()},
\code{\link{LSAMonSetH}()},
\code{\link{LSAMonSetTopicColumnPositions}()},
\code{\link{LSAMonSetTopicNames}()},
\code{\link{LSAMonSetValue}()},
\code{\link{LSAMonSetWeightedDocumentTermMatrix}()},
\code{\link{LSAMonSetW}()},
\code{\link{LSAMonTakeDocumentTermMatrix}()},
\code{\link{LSAMonTakeDocuments}()},
\code{\link{LSAMonTakeGlobalWeights}()},
\code{\link{LSAMonTakeH}()},
\code{\link{LSAMonTakeLocalWeightFunction}()},
\code{\link{LSAMonTakeMethod}()},
\code{\link{LSAMonTakeNormalizerFunction}()},
\code{\link{LSAMonTakeTerms}()},
\code{\link{LSAMonTakeTopicColumnPositions}()},
\code{\link{LSAMonTakeTopicNames}()},
\code{\link{LSAMonTakeValue}()},
\code{\link{LSAMonTakeWeightedDocumentTermMatrix}()},
\code{\link{LSAMonTakeW}()}
}
\concept{Set/Take functions}
