% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonSetTopicColumnPositions}
\alias{LSAMonSetTopicColumnPositions}
\title{Set TopicColumnPositions.}
\usage{
LSAMonSetTopicColumnPositions(lsaObj, topicColumnPositions)
}
\arguments{
\item{lsaObj}{An LSAMon object.}

\item{topicColumnPositions}{A list of integer vectors.}
}
\value{
An LSAMon object.
}
\description{
Sets TopicColumnPositions into the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{LSAMonSetDocumentTermMatrix}},
  \code{\link{LSAMonSetDocuments}},
  \code{\link{LSAMonSetH}}, \code{\link{LSAMonSetTerms}},
  \code{\link{LSAMonSetTopicNames}},
  \code{\link{LSAMonSetValue}},
  \code{\link{LSAMonSetWeightedDocumentTermMatrix}},
  \code{\link{LSAMonSetW}},
  \code{\link{LSAMonTakeDocumentTermMatrix}},
  \code{\link{LSAMonTakeDocuments}},
  \code{\link{LSAMonTakeH}}, \code{\link{LSAMonTakeTerms}},
  \code{\link{LSAMonTakeTopicColumnPositions}},
  \code{\link{LSAMonTakeTopicNames}},
  \code{\link{LSAMonTakeValue}},
  \code{\link{LSAMonTakeWeightedDocumentTermMatrix}},
  \code{\link{LSAMonTakeW}}
}
\concept{Set/Take functions}
