% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonSetTopicNames}
\alias{LSAMonSetTopicNames}
\title{Set AutomaticTopicNames.}
\usage{
LSAMonSetTopicNames(lsaObj, topicNames)
}
\arguments{
\item{lsaObj}{An LSAMon object.}

\item{topicNames}{A list of topic names.}
}
\value{
An LSAMon object.
}
\description{
Sets AutomaticTopicNames into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{LSAMonSetDocumentTermMatrix}()},
\code{\link{LSAMonSetDocuments}()},
\code{\link{LSAMonSetH}()},
\code{\link{LSAMonSetTerms}()},
\code{\link{LSAMonSetTopicColumnPositions}()},
\code{\link{LSAMonSetValue}()},
\code{\link{LSAMonSetWeightedDocumentTermMatrix}()},
\code{\link{LSAMonSetW}()},
\code{\link{LSAMonTakeDocumentTermMatrix}()},
\code{\link{LSAMonTakeDocuments}()},
\code{\link{LSAMonTakeH}()},
\code{\link{LSAMonTakeTerms}()},
\code{\link{LSAMonTakeTopicColumnPositions}()},
\code{\link{LSAMonTakeTopicNames}()},
\code{\link{LSAMonTakeValue}()},
\code{\link{LSAMonTakeWeightedDocumentTermMatrix}()},
\code{\link{LSAMonTakeW}()}
}
\concept{Set/Take functions}
