% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonSetValue}
\alias{LSAMonSetValue}
\title{Set the value in a LSAMon object.}
\usage{
LSAMonSetValue(lsaObj, value)
}
\arguments{
\item{lsaObj}{An LSAMon object.}

\item{value}{The new value.}
}
\value{
A LSAMon object.
}
\description{
Sets the value in a LSAMon monad object.
}
\details{
Assigns \code{value} to \code{lsaObj$Value}.
}
\seealso{
Other Set/Take functions: 
\code{\link{LSAMonSetDocumentTermMatrix}()},
\code{\link{LSAMonSetDocuments}()},
\code{\link{LSAMonSetH}()},
\code{\link{LSAMonSetTerms}()},
\code{\link{LSAMonSetTopicColumnPositions}()},
\code{\link{LSAMonSetTopicNames}()},
\code{\link{LSAMonSetWeightedDocumentTermMatrix}()},
\code{\link{LSAMonSetW}()},
\code{\link{LSAMonTakeDocumentTermMatrix}()},
\code{\link{LSAMonTakeDocuments}()},
\code{\link{LSAMonTakeGlobalWeights}()},
\code{\link{LSAMonTakeH}()},
\code{\link{LSAMonTakeLocalWeightFunction}()},
\code{\link{LSAMonTakeMethod}()},
\code{\link{LSAMonTakeNormalizerFunction}()},
\code{\link{LSAMonTakeTerms}()},
\code{\link{LSAMonTakeTopicColumnPositions}()},
\code{\link{LSAMonTakeTopicNames}()},
\code{\link{LSAMonTakeValue}()},
\code{\link{LSAMonTakeWeightedDocumentTermMatrix}()},
\code{\link{LSAMonTakeW}()}
}
\concept{Set/Take functions}
