% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonSetWeightedDocumentTermMatrix}
\alias{LSAMonSetWeightedDocumentTermMatrix}
\title{Set WeightedDocumentTermMatrix.}
\usage{
LSAMonSetWeightedDocumentTermMatrix(lsaObj, WeightedDocumentTermMatrix)
}
\arguments{
\item{lsaObj}{An LSAMon object.}

\item{WeightedDocumentTermMatrix}{A list of regression objects.}
}
\value{
An LSAMon object.
}
\description{
Sets WeightedDocumentTermMatrix into the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{LSAMonSetAutomaticTopicNames}},
  \code{\link{LSAMonSetDocumentTermMatrix}},
  \code{\link{LSAMonSetDocuments}},
  \code{\link{LSAMonSetH}}, \code{\link{LSAMonSetTerms}},
  \code{\link{LSAMonSetTopicColumnPositions}},
  \code{\link{LSAMonSetValue}}, \code{\link{LSAMonSetW}},
  \code{\link{LSAMonTakeAutomaticTopicNames}},
  \code{\link{LSAMonTakeDocumentTermMatrix}},
  \code{\link{LSAMonTakeDocuments}},
  \code{\link{LSAMonTakeH}}, \code{\link{LSAMonTakeTerms}},
  \code{\link{LSAMonTakeTopicColumnPositions}},
  \code{\link{LSAMonTakeValue}},
  \code{\link{LSAMonTakeWeightedDocumentTermMatrix}},
  \code{\link{LSAMonTakeW}}
}
\concept{Set/Take functions}
