% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonTakeDocumentTermMatrix}
\alias{LSAMonTakeDocumentTermMatrix}
\title{Take DocumentTermMatrix.}
\usage{
LSAMonTakeDocumentTermMatrix(
  lsaObj,
  functionName = "LSAMonTakeDocumentTermMatrix"
)
}
\arguments{
\item{lsaObj}{An LSAMon object.}

\item{functionName}{A string that is a name of this function or a delegating function.}
}
\value{
A list of functions or \code{LSAMonFailureSymbol}.
}
\description{
Takes DocumentTermMatrix from the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{LSAMonSetDocumentTermMatrix}()},
\code{\link{LSAMonSetDocuments}()},
\code{\link{LSAMonSetH}()},
\code{\link{LSAMonSetTerms}()},
\code{\link{LSAMonSetTopicColumnPositions}()},
\code{\link{LSAMonSetTopicNames}()},
\code{\link{LSAMonSetValue}()},
\code{\link{LSAMonSetWeightedDocumentTermMatrix}()},
\code{\link{LSAMonSetW}()},
\code{\link{LSAMonTakeDocuments}()},
\code{\link{LSAMonTakeGlobalWeights}()},
\code{\link{LSAMonTakeH}()},
\code{\link{LSAMonTakeLocalWeightFunction}()},
\code{\link{LSAMonTakeMethod}()},
\code{\link{LSAMonTakeNormalizerFunction}()},
\code{\link{LSAMonTakeTerms}()},
\code{\link{LSAMonTakeTopicColumnPositions}()},
\code{\link{LSAMonTakeTopicNames}()},
\code{\link{LSAMonTakeValue}()},
\code{\link{LSAMonTakeWeightedDocumentTermMatrix}()},
\code{\link{LSAMonTakeW}()}
}
\concept{Set/Take functions}
