% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonTakeNormalizedMatrixProductComponents}
\alias{LSAMonTakeNormalizedMatrixProductComponents}
\title{Take normalized matrix product components.}
\usage{
LSAMonTakeNormalizedMatrixProductComponents(
  lsaObj,
  normalizeLeftQ = TRUE,
  orderBySignificanceQ = FALSE
)
}
\arguments{
\item{lsaObj}{A LSAMon object.}

\item{normalizeLeftQ}{Should the left factor be normalized?}

\item{orderBySignificanceQ}{Should the basis vectors be ordered by their significance?}
}
\value{
A list of matrices
}
\description{
Normalizes the matrix factors and returns a list with them.
}
\details{
Makes a call to \code{\link{LSAMonNormalizeMatrixProduct}}.
}
