% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonTakeStatisticalThesaurus}
\alias{LSAMonTakeStatisticalThesaurus}
\title{Take statistical thesaurus table.}
\usage{
LSAMonTakeStatisticalThesaurus(
  lsaObj,
  words,
  numberOfNearestNeighbors = 12,
  wideFormQ = FALSE
)
}
\arguments{
\item{lsaObj}{A LSAMon object.}

\item{words}{Words to find thesaurus entries for.}

\item{numberOfNearestNeighbors}{Number of nearest neighbors per specified word.}

\item{wideFormQ}{Should the thesaurus table be in wide form or not?}
}
\value{
A LSAMon object.
}
\description{
Takes a table of statistical thesaurus entries for specified words.
}
\details{
This function is a shortcut of \code{\link{LSAMonGetStatisticalThesaurus}}.
}
