% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonTakeWeightedDocumentTermMatrix}
\alias{LSAMonTakeWeightedDocumentTermMatrix}
\title{Take WeightedDocumentTermMatrix.}
\usage{
LSAMonTakeWeightedDocumentTermMatrix(
  lsaObj,
  functionName = "LSAMonTakeWeightedDocumentTermMatrix"
)
}
\arguments{
\item{lsaObj}{An LSAMon object.}

\item{functionName}{A string that is a name of this function or a delegating function.}
}
\value{
A list of functions or \code{LSAMonFailureSymbol}.
}
\description{
Takes WeightedDocumentTermMatrix from the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{LSAMonSetDocumentTermMatrix}()},
\code{\link{LSAMonSetDocuments}()},
\code{\link{LSAMonSetH}()},
\code{\link{LSAMonSetTerms}()},
\code{\link{LSAMonSetTopicColumnPositions}()},
\code{\link{LSAMonSetTopicNames}()},
\code{\link{LSAMonSetValue}()},
\code{\link{LSAMonSetWeightedDocumentTermMatrix}()},
\code{\link{LSAMonSetW}()},
\code{\link{LSAMonTakeDocumentTermMatrix}()},
\code{\link{LSAMonTakeDocuments}()},
\code{\link{LSAMonTakeH}()},
\code{\link{LSAMonTakeTerms}()},
\code{\link{LSAMonTakeTopicColumnPositions}()},
\code{\link{LSAMonTakeTopicNames}()},
\code{\link{LSAMonTakeValue}()},
\code{\link{LSAMonTakeW}()}
}
\concept{Set/Take functions}
