% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonTopicExtraction}
\alias{LSAMonTopicExtraction}
\title{Topic extraction.}
\usage{
LSAMonTopicExtraction(lsaObj, numberOfTopics,
  minNumberOfDocumentsPerTerm = NULL, maxSteps = 12,
  tolerance = 0.01, profiling = TRUE)
}
\arguments{
\item{lsaObj}{A LSAMon object.}

\item{numberOfTopics}{Number of topics to be extracted.}

\item{minNumberOfDocumentsPerTerm}{Minimal number of documents for the terms
to be considered in the topics.}

\item{maxSteps}{Maximum iteration steps.}
}
\value{
A LSAMon object.
}
\description{
Topic extraction from a document-term matrix using
Non Negative Matrix Factorization (NNMF).
}
\details{
The obtained factor matrices are assigned to \code{lsaObj$W} and \code{lsaObj$H}.
The parameters \code{maxSteps, tolerance, profiling} are
for the function \code{\link{NonNegativeMatrixFactorization::NNMF}}.
If \code{is.null(minNumberOfDocumentsPerTerm)} then the following formula is used:
\code{minNumberOfDocumentsPerTerm = floor(0.05*nrow(lsaObj$DocumentTermMatrix))}.
}
