% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonTopicRepresentation}
\alias{LSAMonTopicRepresentation}
\title{Topic representation.}
\usage{
LSAMonTopicRepresentation(lsaObj, tags = NULL, minThreshold = 0.001)
}
\arguments{
\item{lsaObj}{A LSAMon object.}

\item{tags}{A character vector with tags that correspond to the documents.
If NULL the documents ordinal numbers or ID's are used.}

\item{minThreshold}{The minimum topic weight.}
}
\value{
The result is a tag-topic contingency matrix that is assigned to
\code{lsaObj$Value}.
}
\description{
Find the topic representation corresponding to a list of tags.
Each monad document is expected to have a tag.
One tag might correspond to multiple documents.
}
