% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{MostImportantSentences}
\alias{MostImportantSentences}
\title{Most important sentences.}
\usage{
MostImportantSentences(
  sentences,
  nTopSentences = 5,
  globalTermWeightFunction = "IDF",
  splitPattern = "\\\\W",
  stemWordsQ = FALSE,
  minWordLength = 2,
  stopWords = NULL
)
}
\arguments{
\item{sentences}{A character vector with sentences.}

\item{nTopSentences}{The number of top most important sentences
to be returned.}

\item{globalTermWeightFunction}{The global term weight function to applied.}

\item{splitPattern}{A string with a regex split pattern.
With that pattern the sentences are split into words.}

\item{stemWordsQ}{Should word stemming be applied or not?}

\item{minWordLength}{Minimal word length.}

\item{stopWords}{A character vector with stop words to be removed.}
}
\value{
A data frame
}
\description{
Finds the most important sentences in a list of sentences.
}
\details{
The result is a data frame with columns \code{c("Score", "Sentence")}.
This is a stand-alone, non-monadic version of \code{\link{LSAMonFindMostImportantDocuments}}.
}
