% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonMakeDocumentTermMatrix}
\alias{LSAMonMakeDocumentTermMatrix}
\title{Document-term matrix creation.}
\usage{
LSAMonMakeDocumentTermMatrix(
  lsaObj,
  splitPattern = "[[:space:]]|[[:punct:]]",
  stemWordsQ = FALSE,
  stemRules = NULL,
  stopWords = NULL
)
}
\arguments{
\item{lsaObj}{A LSAMon object.}

\item{splitPattern}{Pattern to use for splitting the documents into words.
If NULL \code{"[[:space:]]|[[:punct:]]"} is used.}

\item{stemWordsQ}{Should the words be stemmed or not?
If \code{stemWordsQ = TRUE} and \code{stemRules = NULL}
then \code{\link{SnowballC::wordStem}} is used.}

\item{stemRules}{A character vector with named elements or NULL.
(A vector of "stemming rules".)
If \code{stemWordsQ = FALSE} then \code{stemRules} is ignored.}

\item{stopWords}{A character vector with stop words to be removed.}
}
\value{
A LSAMon object.
}
\description{
Make a document-term matrix for a collection of documents
using specified word splitting and optional word stemming.
}
\details{
The obtained sparse matrix is \code{lsaObj$DocumentTermMatrix}.
The documents are expected to have ID's. If \code{is.null(names(lsaObj$Documents))}
then the ID's are just the indexes of the documents list/vector.
Stop words can be supplied with \code{\link{stopwords::stopwords}}.
}
