% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonSetDocumentTermMatrix}
\alias{LSAMonSetDocumentTermMatrix}
\title{Set DocumentTermMatrix.}
\usage{
LSAMonSetDocumentTermMatrix(lsaObj, DocumentTermMatrix)
}
\arguments{
\item{lsaObj}{An LSAMon object.}

\item{documentTermMatrix}{A document term sparse matrix.
(Of type \code{dgCMatrix}.)}
}
\value{
An LSAMon object.
}
\description{
Sets DocumentTermMatrix into the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{LSAMonSetDocuments}},
  \code{\link{LSAMonSetH}}, \code{\link{LSAMonSetTerms}},
  \code{\link{LSAMonSetTopicColumnPositions}},
  \code{\link{LSAMonSetTopicNames}},
  \code{\link{LSAMonSetValue}},
  \code{\link{LSAMonSetWeightedDocumentTermMatrix}},
  \code{\link{LSAMonSetW}},
  \code{\link{LSAMonTakeDocumentTermMatrix}},
  \code{\link{LSAMonTakeDocuments}},
  \code{\link{LSAMonTakeH}}, \code{\link{LSAMonTakeTerms}},
  \code{\link{LSAMonTakeTopicColumnPositions}},
  \code{\link{LSAMonTakeTopicNames}},
  \code{\link{LSAMonTakeValue}},
  \code{\link{LSAMonTakeWeightedDocumentTermMatrix}},
  \code{\link{LSAMonTakeW}}
}
\concept{Set/Take functions}
