% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonSetWeightedDocumentTermMatrix}
\alias{LSAMonSetWeightedDocumentTermMatrix}
\title{Set WeightedDocumentTermMatrix.}
\usage{
LSAMonSetWeightedDocumentTermMatrix(lsaObj, weightedDocumentTermMatrix)
}
\arguments{
\item{lsaObj}{An LSAMon object.}

\item{weightedDocumentTermMatrix}{A weighted document term sparse matrix.
(Of type \code{dgCMatrix}.)}
}
\value{
An LSAMon object.
}
\description{
Sets WeightedDocumentTermMatrix into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{LSAMonSetDocumentTermMatrix}()},
\code{\link{LSAMonSetDocuments}()},
\code{\link{LSAMonSetH}()},
\code{\link{LSAMonSetTerms}()},
\code{\link{LSAMonSetTopicColumnPositions}()},
\code{\link{LSAMonSetTopicNames}()},
\code{\link{LSAMonSetValue}()},
\code{\link{LSAMonSetW}()},
\code{\link{LSAMonTakeDocumentTermMatrix}()},
\code{\link{LSAMonTakeDocuments}()},
\code{\link{LSAMonTakeH}()},
\code{\link{LSAMonTakeTerms}()},
\code{\link{LSAMonTakeTopicColumnPositions}()},
\code{\link{LSAMonTakeTopicNames}()},
\code{\link{LSAMonTakeValue}()},
\code{\link{LSAMonTakeWeightedDocumentTermMatrix}()},
\code{\link{LSAMonTakeW}()}
}
\concept{Set/Take functions}
