% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonTopicExtraction}
\alias{LSAMonTopicExtraction}
\title{Topic extraction.}
\usage{
LSAMonTopicExtraction(lsaObj, numberOfTopics,
  minNumberOfDocumentsPerTerm = NULL, method = "NNMF",
  maxSteps = NULL, tolerance = 0.01, profilingQ = TRUE,
  orderBySignificanceQ = TRUE, automaticTopicNamesQ = TRUE, ...)
}
\arguments{
\item{lsaObj}{A LSAMon object.}

\item{numberOfTopics}{Number of topics to be extracted.}

\item{minNumberOfDocumentsPerTerm}{Minimal number of documents for the terms
to be considered in the topics.}

\item{maxSteps}{Maximum iteration steps.
If NULL appropriate default values for the different methods are used.
(For NNMF 12, for SVD 1000.)}

\item{tolerance}{Tolerance for the relative norm difference.}

\item{profilingQ}{Should the computation be profiled?}

\item{orderBySignificanceQ}{Should the basis vectors be ordered by their significance?}

\item{automaticTopicNamesQ}{Should the extracted topics be given automatic names?}

\item{...}{Additional parameters for the matrix decomposition functions.
See \code{\link{irlba::irlba}} and \code{\link{NonNegativeMatrixFactorization::NNMF}}.
Note that some of those overlap with some of function's arguments.}
}
\value{
A LSAMon object.
}
\description{
Topic extraction from a document-term matrix using
Non Negative Matrix Factorization (NNMF).
}
\details{
The obtained factor matrices are assigned to \code{lsaObj$W} and \code{lsaObj$H}.
The parameters \code{maxSteps, tolerance, profilingQ} are
for the function \code{\link{NonNegativeMatrixFactorization::NNMF}}.
If \code{is.null(minNumberOfDocumentsPerTerm)} then the following formula is used:
\code{minNumberOfDocumentsPerTerm = floor(0.05*nrow(lsaObj$DocumentTermMatrix))}.
The automatic topic name is derived from topic's index and topic's top three words.
(The later is to ensure topic name uniqueness.)
}
