% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonTopicRepresentation}
\alias{LSAMonTopicRepresentation}
\title{Topic representation.}
\usage{
LSAMonTopicRepresentation(lsaObj, tags = NULL, minThreshold = 0.001,
  normalizeLeftQ = TRUE)
}
\arguments{
\item{lsaObj}{A LSAMon object.}

\item{tags}{A character vector with tags that correspond to the documents.
The correspondence is ordinal or through the names of \code{tags}.
If NULL the documents ordinal numbers or ID's are used.}

\item{minThreshold}{The minimum absolute topic weight.}

\item{normalizeLeftQ}{Should the left matrix multiplication factor be normalized?}
}
\value{
The result is a tag-topic contingency matrix that is assigned to
\code{lsaObj$Value}.
}
\description{
Find the topic representation corresponding to a list of tags.
Each monad document is expected to have a tag.
One tag might correspond to multiple documents.
}
\details{
Monad's document collection is given as a character vector \code{texts}.
If \code{texts} does not have element names, i.e. \code{names(texts) == NULL},
they are assigned to be the corresponding indexes.
Those element names of \code{texts} are used as document ID's.
The document ID's are used if \code{tags = NULL} is specified.
When this function is called, it is expected that the dimension reduction is
already computed. Hence, the rownames of the left matrix factor \code{lsaObj$W}
are expected to correspond to the document ID's.
The argument \code{tags} itself is a mapping vector: \code{names(tags)} is
supposed to correspond to the document ID's and the values of \code{tags} are
the tags to be represented with topics.
Using \code{tags=NULL} should produce a matrix derived from matrix
\code{lsaObj$W} by removing all entries from \code{lsaObj$W} that have
absolute values less than \code{minThreshold}.
Absolute values have to be considered for dimension reduction
algorithms as SVD and ICA.
}
