% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAMon.R
\name{LSAMonUnit}
\alias{LSAMonUnit}
\title{Make a LSAMon Unit}
\usage{
LSAMonUnit()
}
\arguments{
\item{data}{A vector or a two-column matrix or data frame.}
}
\value{
An S3 class "LSAMon". In other words, a list with the attribute "class" set to "LSAMon".
}
\description{
Creates a monad object.
}
